corrected imports in discretizers
