+++ AutoCarver/BaseDiscretizers.py
@@ -10,8 +10,6 @@ from pandas import DataFrame, Series, isna, notna, unique
 from pandas.api.types import is_numeric_dtype, is_string_dtype
 from sklearn.base import BaseEstimator, TransformerMixin
 
-from .Converters import StringConverter
-
 
 def nan_unique(x: Series):
     """Unique non-NaN values."""
@@ -519,3 +517,167 @@ class ClosestDiscretizer(BaseEstimator, TransformerMixin):
             Xc[feature] = select(to_input, order, default=Xc[feature])  # regroupement des valeurs
 
         return Xc
+
+
+def find_common_modalities(
+    df_feature: Series,
+    y: Series,
+    min_freq: float,
+    order: GroupedList,
+    len_df: int = None,
+) -> Dict[str, Any]:
+    """[Qualitative] Découpage en modalités de fréquence minimal (Cas des variables ordonnées).
+
+    Fonction récursive : on prend l'échantillon et on cherche des valeurs sur-représentées
+    Si la valeur existe on la met dans une classe et on cherche à gauche et à droite de celle-ci, d'autres variables sur-représentées
+    Une fois il n'y a plus de valeurs qui soient sur-représentées,
+    on fait un découpage classique avec qcut en multipliant le nombre de classes souhaité par le pourcentage de valeurs restantes sur le total
+
+    """
+
+    # initialisation de la taille totale du dataframe
+    if len_df is None:
+        len_df = len(df_feature)
+
+    # conversion en GroupedList si ce n'est pas le cas
+    order = GroupedList(order)
+
+    # cas 1 : il y a des valeurs manquantes NaN
+    if any(isna(df_feature)):
+        return find_common_modalities(
+            df_feature[notna(df_feature)],
+            y[notna(df_feature)],
+            min_freq,
+            order,
+            len_df,
+        )
+
+    # cas 2 : il n'y a que des valeurs dans le dataframe (pas de NaN)
+    else:
+        # computing frequencies and target rate of each modality
+        init_frequencies = df_feature.value_counts(dropna=False, normalize=False) / len_df
+        init_values, init_frequencies = (
+            init_frequencies.index,
+            init_frequencies.values,
+        )
+
+        # ordering
+        frequencies = [
+            init_frequencies[init_values == value][0] if any(init_values == value) else 0
+            for value in order
+        ]  # sort selon l'ordre des modalités
+        values = [
+            init_values[init_values == value][0] if any(init_values == value) else value
+            for value in order
+        ]  # sort selon l'ordre des modalités
+        target_rate = (
+            y.groupby(df_feature).sum().reindex(order) / frequencies / len_df
+        )  # target rate per modality
+        underrepresented = [
+            value for value, frequency in zip(values, frequencies) if frequency < min_freq
+        ]  # valeur peu fréquentes
+
+        # cas 1 : il existe une valeur sous-représentée
+        while any(underrepresented) & (len(frequencies) > 1):
+            # identification de la valeur sous-représentée
+            discarded_idx = argmin(frequencies)
+            discarded = values[discarded_idx]
+
+            # identification de la modalité la plus proche (volume et taux de défaut)
+            kept = find_closest_modality(
+                discarded,
+                discarded_idx,
+                list(zip(order, frequencies, target_rate)),
+                min_freq,
+            )
+
+            # removing the value from the initial ordering
+            order.group(discarded, kept)
+
+            # ordering
+            frequencies = [
+                init_frequencies[init_values == value][0] if any(init_values == value) else 0
+                for value in order
+            ]  # sort selon l'ordre des modalités
+            values = [
+                init_values[init_values == value][0] if any(init_values == value) else value
+                for value in order
+            ]  # sort selon l'ordre des modalités
+            target_rate = (
+                y.groupby(df_feature).sum().reindex(order) / frequencies / len_df
+            )  # target rate per modality
+            underrepresented = [
+                value for value, frequency in zip(values, frequencies) if frequency < min_freq
+            ]  # valeur peu fréquentes
+
+        worst, best = target_rate.idxmin(), target_rate.idxmax()
+
+        # cas 2 : il n'existe pas de valeur sous-représentée
+        return {"order": order, "worst": worst, "best": best}
+
+
+def find_closest_modality(value, idx: int, freq_target: Series, min_freq: float) -> int:
+    """[Qualitative] HELPER Finds the closest modality in terms of frequency and target rate"""
+
+    # case 1: lowest modality
+    if idx == 0:
+        closest_modality = 1
+
+    # case 2: biggest modality
+    elif idx == len(freq_target) - 1:
+        closest_modality = len(freq_target) - 2
+
+    # case 3: not the lowwest nor the biggest modality
+    else:
+        # previous modality's volume and target rate
+        previous_value, previous_volume, previous_target = freq_target[idx - 1]
+
+        # current modality's volume and target rate
+        _, volume, target = freq_target[idx]
+
+        # next modality's volume and target rate
+        next_value, next_volume, next_target = freq_target[idx + 1]
+
+        # regroupement par volumétrie (préféré s'il n'y a pas beaucoup de volume)
+        # case 1: la modalité suivante est inférieure à min_freq
+        if next_volume < min_freq <= previous_volume:
+            closest_modality = idx + 1
+
+        # case 2: la modalité précédante est inférieure à min_freq
+        elif previous_volume < min_freq <= next_volume:
+            closest_modality = idx - 1
+
+        # case 3: les deux modalités sont inférieures à min_freq
+        elif (previous_volume < min_freq) & (next_volume < min_freq):
+            # cas 1: la modalité précédante est inférieure à la modalité suivante
+            if previous_volume < next_volume:
+                closest_modality = idx - 1
+
+            # cas 2: la modalité suivante est inférieure à la modalité précédante
+            elif next_volume < previous_volume:
+                closest_modality = idx + 1
+
+            # cas 3: les deux modalités ont la même fréquence
+            else:
+                # cas1: la modalité précédante est plus prorche en taux de cible
+                if abs(previous_target - target) <= abs(next_target - target):
+                    closest_modality = idx - 1
+
+                # cas2: la modalité précédente est plus proche en taux de cible
+                else:
+                    closest_modality = idx + 1
+
+        # case 4: les deux modalités sont supérieures à min_freq
+        else:
+            # cas1: la modalité précédante est plus prorche en taux de cible
+            if abs(previous_target - target) <= abs(next_target - target):
+                closest_modality = idx - 1
+
+            # cas2: la modalité précédante est plus prorche en taux de cible
+            else:
+                closest_modality = idx + 1
+
+    # récupération de la valeur associée
+    closest_value = freq_target[closest_modality][0]
+
+    return closest_value
